+++ src/controllers/ElementsController.php
@@ -1059,6 +1059,17 @@ JS, [
                             }
                         }
 
+                        // If the error is for a recursively-nested Matrix field,
+                        // manipulate the key to only reference the nested Matrix field, entry and inner field
+                        // Before: foo[<uuid>].bar[<uuid>].baz
+                        // After:  bar[<uuid>].baz
+                        if (substr_count($key, '.') > 1) {
+                            $keyParts = explode('.', $key);
+                            if (preg_match(sprintf('/\[%s\]$/', StringHelper::UUID_PATTERN), $keyParts[count($keyParts) - 3])) {
+                                $key = implode('.', array_slice($keyParts, -2));
+                            }
+                        }
+
                         $errorItem = null;
                         if ($error !== null) {
                             $error = Markdown::processParagraph(htmlspecialchars($error));